requirements.txt
pandas
numpy
scikit-learn
imbalanced-learn
category-encoders
shap
lime
matplotlib
plotly
python-dotenv
pyyaml
pyarrow
mlflow
streamlit
fastapi
uvicorn
pytest
black
isort
ruff
pre-commit
